+++ src/Core/Generator/OperationBuilder.php
@@ -524,7 +524,7 @@ final readonly class OperationBuilder
                 ...$descriptor->actionAttributes(TagAttribute::class),
             ] as $attr
         ) {
-            $tags[] = $attr->newInstance()->name;
+            $tags[] = $attr->newInstance()->value();
         }
 
         return $tags;